feat: remove Embroider compat mode from performance-test-app
Drops the classic ember-cli/broccoli "stage1" pass this app kept
around solely for Embroider's compat build path: removes
ember-cli-build.js, classicEmberSupport() from vite.config.mjs,
@embroider/compat/babel's compat helpers from the babel config, and
the classic app.js/router.js + config/environment.js <-> app/config
/environment.js (config-meta-loader) bridge.

Flattens app/{models,mixins,routes,services,styles,templates} to
root-level directories (this app has no other reason to keep an app/
prefix once ember-cli's conventions no longer apply) and replaces the
classic Resolver.withModules(compatModules) app boot with
ember-strict-application-resolver, wiring the flattened directories in
via import.meta.glob(...) spread into the app's `modules` map --
matching the pattern tests/example-app-ember already established for
a real (non-diagnostic) native Embroider+Vite app. Also required
adding the `hbs()` plugin from @embroider/vite to vite.config.mjs,
which the compat build's classicEmberSupport() had been providing
raw-.hbs-template handling for.

This app has no automated test suite (it's a TracerBench benchmark
build target) -- verified by building, serving the dist output with
the app's own server/index.ts, and loading several representative
routes (including one exercising all 10 polymorphic mixin models)
headlessly, confirming no console errors and correct rendering/routing.
