--- tests/performance/app/app.js
@@ -1,23 +1,0 @@
-// uncomment this to install Map/Set instrumentation
-// prior to app boot
-// import './services/store';
-import '@warp-drive/ember/install';
-
-import Application from '@ember/application';
-
-import compatModules from '@embroider/virtual/compat-modules';
-
-import loadInitializers from 'ember-load-initializers';
-import Resolver from 'ember-resolver';
-
-import config from './config/environment';
-
-class App extends Application {
-  modulePrefix = config.modulePrefix;
-  podModulePrefix = config.podModulePrefix;
-  Resolver = Resolver.withModules(compatModules);
-}
-
-loadInitializers(App, config.modulePrefix, compatModules);
-
-export default App;
