--- tests/performance/app/config/environment.js
@@ -1,3 +1,0 @@
-import loadConfigFromMeta from '@embroider/config-meta-loader';
-
-export default loadConfigFromMeta('performance-test-app');
