--- tests/performance/babel.config.cjs
@@ -1,46 +1,0 @@
-const { babelCompatSupport, templateCompatSupport } = require('@embroider/compat/babel');
-
-module.exports = {
-  plugins: [
-    [
-      '@babel/plugin-transform-typescript',
-      {
-        allExtensions: true,
-        onlyRemoveTypeImports: true,
-        allowDeclareFields: true,
-      },
-    ],
-    [
-      'babel-plugin-ember-template-compilation',
-      {
-        enableLegacyModules: [
-          'ember-cli-htmlbars',
-          'ember-cli-htmlbars-inline-precompile',
-          'htmlbars-inline-precompile',
-        ],
-        transforms: [...templateCompatSupport()],
-      },
-    ],
-    [
-      'module:decorator-transforms',
-      {
-        runtime: {
-          import: require.resolve('decorator-transforms/runtime-esm'),
-        },
-      },
-    ],
-    [
-      '@babel/plugin-transform-runtime',
-      {
-        absoluteRuntime: __dirname,
-        useESModules: true,
-        regenerator: false,
-      },
-    ],
-    ...babelCompatSupport(),
-  ],
-
-  generatorOpts: {
-    compact: false,
-  },
-};
