--- tests/performance/config/optional-features.json
@@ -1,7 +1,0 @@
-{
-  "application-template-wrapper": false,
-  "default-async-observers": true,
-  "jquery-integration": false,
-  "template-only-glimmer-components": true,
-  "no-implicit-route-model": true
-}
