--- tests/performance/config/targets.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const browsers = ['last 1 Chrome versions', 'last 1 Firefox versions', 'last 1 Safari versions'];
-
-module.exports = {
-  browsers,
-  node: 'current',
-};
