+++ tests/performance/index.html
@@ -6,25 +6,13 @@
     <title>RelationshipPerformance</title>
     <meta name="description" content="" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-    {{content-for "head"}}
-
-    <link integrity="" rel="stylesheet" href="/@embroider/virtual/vendor.css">
-    <link integrity="" rel="stylesheet" href="/@embroider/virtual/app.css">
-
-    {{content-for "head-footer"}}
   </head>
   <body>
-    {{content-for "body"}}
-
-    <script type="module" src="/@embroider/virtual/vendor.js"></script>
     <script type="module">
-      import Application from './app/app';
-      import environment from './app/config/environment';
+      import Application from './app.ts';
+      import './styles/app.css';
 
-      Application.create(environment.APP);
+      Application.create();
     </script>
-
-    {{content-for "body-footer"}}
   </body>
 </html>
