+++ tests/performance/mixins/record-mixin-b.js
