+++ tests/performance/mixins/record-mixin-c.js
