+++ tests/performance/mixins/record-mixin-f.js
