+++ tests/performance/package.json
@@ -23,7 +23,6 @@
   "devDependencies": {
     "@babel/core": "^7.28.3",
     "@babel/runtime": "^7.28.3",
-    "@ember/optional-features": "^2.2.0",
     "@babel/plugin-transform-runtime": "^7.28.3",
     "@babel/plugin-transform-typescript": "^7.28.0",
     "@ember/test-helpers": "5.2.0",
@@ -34,23 +33,17 @@
     "@warp-drive/json-api": "workspace:*",
     "@warp-drive/utilities": "workspace:*",
     "@warp-drive/legacy": "workspace:*",
-    "@embroider/compat": "4.1.12",
-    "@embroider/config-meta-loader": "1.0.0",
     "@embroider/core": "^4.4.2",
+    "@embroider/macros": "^1.20.6",
     "@embroider/vite": "^1.5.0",
     "@glimmer/component": "^2.0.0",
     "@rollup/plugin-babel": "^6.0.4",
     "babel-plugin-ember-template-compilation": "^4.0.0",
     "bun-types": "^1.4.0",
     "decorator-transforms": "^2.3.0",
-    "ember-auto-import": "^2.13.1",
-    "ember-cli": "~6.12.0",
-    "ember-cli-babel": "^8.2.0",
-    "ember-cli-htmlbars": "^7.0.1",
-    "ember-load-initializers": "^3.0.1",
-    "ember-resolver": "^13.1.1",
     "ember-source": "~6.12.0",
-    "loader.js": "^4.7.0",
+    "ember-strict-application-resolver": "^0.1.0",
+    "rollup": "^4.48.0",
     "terser": "^5.43.1",
     "vite": "^7.3.1",
     "vite-bundle-analyzer": "^1.2.1",
@@ -65,12 +58,5 @@
   },
   "volta": {
     "extends": "../../package.json"
-  },
-  "ember-addon": {
-    "type": "app",
-    "version": 2
-  },
-  "exports": {
-    "./*": "./app/*"
   }
 }
