+++ tests/performance/routes/add-children-then-materialize.js
