+++ tests/performance/routes/destroy.js
