--- tests/performance/tests/index.html
@@ -1,1 +1,0 @@
-<html></html>
