+++ tests/performance/vite.config.mjs
@@ -1,5 +1,5 @@
 import { defineConfig } from 'vite';
-import { extensions, classicEmberSupport, ember } from '@embroider/vite';
+import { extensions, ember, hbs } from '@embroider/vite';
 import { babel } from '@rollup/plugin-babel';
 import { compression } from 'vite-plugin-compression2';
 // import { analyzer } from 'vite-bundle-analyzer';
@@ -8,7 +8,7 @@ import zlib from 'zlib';
 
 export default defineConfig({
   plugins: [
-    classicEmberSupport(),
+    hbs(),
     ember(),
     // extra plugins here
     babel({
